gpu-worker/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
requests==2.31.0
numpy==1.26.3
Pillow==10.2.0
# PyTorch with CUDA 11.8 support
torch==2.2.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.17.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
# AI Libraries
rembg[gpu]==2.0.56
onnxruntime-gpu==1.17.0
# Real-ESRGAN dependencies
basicsr==1.4.2
facexlib>=0.2.5
gfpgan>=1.3.5
realesrgan==0.3.0
opencv-python-headless==4.9.0.80
